requirements.txt
Django>=3.2,<3.3
mysqlclient
django-debug-toolbar
django-crispy-forms
colorful
django_social_share
django-allauth
pyopenssl
django-paypal
sentry-sdk
django-crontab
celery
redis
django-celery-beat
django-bootstrap-modal-forms
crispy-bootstrap5
django-storages
boto3
PyGithub
Django-filter
Pillow
stripe